+++ ConnectionDAMForms/ConnectionDAMForms/ConnectionDAMForms.csproj
@@ -100,18 +100,27 @@
   <ItemGroup>
     <Compile Include="ClErrors.cs" />
     <Compile Include="ClPort.cs" />
-    <Compile Include="ServerConnection\ClConnection.cs" />
-    <Compile Include="ClSocket.cs" />
     <Compile Include="FrmMain.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="FrmMain.Designer.cs">
       <DependentUpon>FrmMain.cs</DependentUpon>
     </Compile>
+    <Compile Include="ServerConnection\ClConnection.cs" />
+    <Compile Include="ClSocket.cs" />
+    <Compile Include="Helper.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Helper.Designer.cs">
+      <DependentUpon>Helper.cs</DependentUpon>
+    </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="FrmMain.resx">
       <DependentUpon>FrmMain.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Helper.resx">
+      <DependentUpon>Helper.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
