+++ ConnectionDAMForms/ConnectionDAMForms/Program.cs
@@ -16,7 +16,7 @@ namespace ConnectionDAMForms
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new FrmMain());
+            Application.Run(new Helper());
         }
     }
 }
